+++ src/main/java/fise2/hpp/ok/eventsIR/Friendship.java
@@ -2,7 +2,7 @@ package fise2.hpp.ok.eventsIR;
 
 public class Friendship implements Event {
 
-    public int ts;
+    public long ts;
     public int user_id_1;
     public int user_id_2;
