+++ src/main/java/fise2/hpp/ok/eventsIR/Like.java
@@ -2,7 +2,7 @@ package fise2.hpp.ok.eventsIR;
 
 public class Like implements Event {
 
-    public int ts;
+    public long ts;
     public int user_id;
     public int comment_id;
