+++ src/main/java/fise2/hpp/ok/eventsIR/Post.java
@@ -2,7 +2,7 @@ package fise2.hpp.ok.eventsIR;
 
 public class Post implements Event {
 
-    public int ts;
+    public long ts;
     public int post_id;
     public int user_id;
     public String post;
